Move shared utils to a common folder
