+++ client/src/components/Notices/NoticeDetail.js
@@ -6,7 +6,8 @@ import {withDataProviders} from 'data-provider'
 import {noticeDetailProvider} from '../../dataProviders/noticesDataProviders'
 import {noticeDetailSelector} from '../../selectors'
 import ExternalLink from '../shared/ExternalLink'
-import {getSuspectLevelLimit, ShowNumberCurrency} from './utilities'
+import {getSuspectLevelLimit} from './utilities'
+import {ShowNumberCurrency} from '../../services/utilities'
 import CompaniesTable from './CompaniesTable'
 import './NoticeDetail.css'
 import NoticeInformation from './NoticeInformation'
